Struct heapfile::HeapFile

pub struct HeapFile { /* private fields */ }

A file in memory that is backed by the heap, i.e., a Vec.

Implementations

impl HeapFile

pub fn create(name: String, parent: &DirRef) -> Result<FileRef, &'static str>

Creates a new file with empty content in the given parent directory. No allocation is performed.

pub fn from_vec(
vec: Vec<u8>,
name: String,
parent: &DirRef
) -> Result<FileRef, &'static str>

Creates a new HeapFile in the given parent directory with the contents of the given Vec. No additional allocation or reallocation is performed.